# [tests] Speed up AnimateDiff test