@a-r-r-o-w 2024-04-22
Hi @sayakpaul @DN6, some updates on the failing tests.

> FAILED tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_from_pipe_consistent_forward_pass_cpu_offload

This fails in both old and new version of tests and I believe it is somewhat device dependent. I tested on different diffusers checkouts but seems like it always fails locally for me but not on CI.

> FAILED tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_from_pipe_consistent_config - RuntimeError: Error(s) in loading state_dict for ModuleList:

This test might need some rework because it assumes that the underlying model is the same as `hf-internal-testing/tiny-stable-diffusion-pipe` but since we've modified UNet to make the tests faster, it no longer can load the state dict correctly.